end-to-end data layer has been prepared for the detail page.

+++ app/src/main/java/com/salihkinali/gameguide/data/api/GameApi.kt
@@ -13,5 +13,5 @@ interface GameApi {
     suspend fun getTotalGameInfo(@Query("key") key: String = BuildConfig.API_KEY): TotalGameResponse
 
     @GET("api/games/{id}/")
-    suspend fun getSingleGameInfo(@Path("id") id: Long, @Query("key") key: String=BuildConfig.API_KEY): SingleGameResponse
+    suspend fun getSingleGameInfo(@Path("id") id: Int, @Query("key") key: String=BuildConfig.API_KEY): SingleGameResponse
 }
