+++ app/src/main/java/com/salihkinali/gameguide/data/source/RemoteDataSource.kt
@@ -6,5 +6,5 @@ import com.salihkinali.gameguide.data.dto.game.Result
 
 interface RemoteDataSource {
     suspend fun getTotalGameInfo(): NetworkResponse<List<Result>>
-    suspend fun getSingleCharacter(id: Long): NetworkResponse<SingleGameResponse>
+    suspend fun getSingleCharacter(id: Int): NetworkResponse<SingleGameResponse>
 }
